requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-query-counter==0.4.0
pillow==11.0.0
psycopg2-binary==2.9.10
setuptools==75.6.0
sqlparse==0.5.2
tabulate==0.9.0
tzdata==2024.2
